+++ platform/ios/xcode/FullStacked/FullStackedApp.swift
@@ -5,6 +5,7 @@ struct FullStackedApp: App {
     var body: some Scene {
         WindowGroup {
             Main()
+                .frame(width: 1200, height: 700)
         }
     }
 }
