+++ platform/ios/xcode/FullStacked/WebView.swift
@@ -44,24 +44,24 @@ class WebView: WKWebView, WKNavigationDelegate, WKScriptMessageHandler, WKDownlo
         fatalError("init(coder:) has not been implemented")
     }
     
-    override var safeAreaInsets: UIEdgeInsets {
-        return UIEdgeInsets(top: super.safeAreaInsets.top, left: 0, bottom: 0, right: 0)
-    }
+//    override var safeAreaInsets: UIEdgeInsets {
+//        return UIEdgeInsets(top: super.safeAreaInsets.top, left: 0, bottom: 0, right: 0)
+//    }
         
-    func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction, decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
-        if(navigationAction.shouldPerformDownload) {
-            decisionHandler(.download)
-        }else if navigationAction.navigationType == .linkActivated  {
-            if let url = navigationAction.request.url, "localhost" != url.host, UIApplication.shared.canOpenURL(url) {
-                UIApplication.shared.open(url)
-                decisionHandler(.cancel)
-            } else {
-                decisionHandler(.allow)
-            }
-        } else {
-            decisionHandler(.allow)
-        }
-    }
+//    func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction, decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
+//        if(navigationAction.shouldPerformDownload) {
+//            decisionHandler(.download)
+//        }else if navigationAction.navigationType == .linkActivated  {
+//            if let url = navigationAction.request.url, "localhost" != url.host, UIApplication.shared.canOpenURL(url) {
+//                UIApplication.shared.open(url)
+//                decisionHandler(.cancel)
+//            } else {
+//                decisionHandler(.allow)
+//            }
+//        } else {
+//            decisionHandler(.allow)
+//        }
+//    }
     
     func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
         if(!self.firstContact){
@@ -109,7 +109,7 @@ class WebView: WKWebView, WKNavigationDelegate, WKScriptMessageHandler, WKDownlo
     }
     
     func downloadDidFinish(_ download: WKDownload) {
-        UIApplication.shared.open(URL(string: "shareddocuments://" + downloadDirectory)!)
+//        UIApplication.shared.open(URL(string: "shareddocuments://" + downloadDirectory)!)
     }
 }
